requirements.txt
transformers==4.38.1
langchain==0.1.6
sentence-transformers==2.5.1
sentencepiece==0.1.99
chromadb==0.4.24
streamlit
streamlit_chat
python-dotenv
rank-bm25==0.2.2